Avoid some compilation warnings

+++ src/drivers/console/stdio.c
@@ -69,5 +69,6 @@ static void _stdio_console_print(STDIOConsole * console, char const * str,
 	size_t i;
 
 	for(i = 0; i < len; i++)
-		data->bus->write8(data->bus, STDIO_ADDRESS_STDOUT, str[i]);
+		data->bus->write8(data->bus, STDIO_ADDRESS_STDOUT,
+				(uint8_t)str[i]);
 }
